requirements.txt
streamlit==1.28.0
python-docx==0.8.11
pandas==2.0.3
matplotlib==3.7.2
openpyxl==3.1.2
reportlab==4.0.4
google-generativeai==0.3.0
plotly==5.15.0
python-dotenv==1.0.0